#1043 P3: relocate NAT show cases to server_show_nat.go
Phase 3 of #1043 takes a ~270 LOC bite out of server_show.go's
modularity-discipline violation by extracting the seven NAT-related
ShowText case bodies (`nat-static`, `nat-nptv6`, `persistent-nat`,
`nat-source-rule-detail`, `nat-dest-rule-detail`,
`persistent-nat-detail`, `nat64`) into a new sibling file
`server_show_nat.go`.

Methodology matches Phases 1 (#1148) and 2 (#1150): semantic relocation,
no behavior change. Each case body is moved verbatim apart from
(a) `&buf` references becoming `buf` (passed-in `*strings.Builder`)
and (b) the original `if … { … } else { … }` structure flattened into
an early-return form (`if … { …; return }; …`) so the extracted methods
read with one fewer indentation level. Output is unchanged.

server_show.go: 3,873 → 3,603 LOC (-270).
server_show_nat.go: new, 338 LOC.

`encoding/binary` import in server_show.go is dropped (only callsite
moved to the new file). All 880+ Go tests still pass.
